Skip projects that are set for retry
If the project failed in the past we skip executing its tasks during a normal run and instead save again the previous config.

We tell the user to fix issues and then attempt rebuilding with "nino -r" or to simply force rebuild with the new config from nino.toml by running "nino -f projectname".